# go/docs/openapi/swagger.yaml
basePath: /
definitions:
...
  entity.Track:
    properties:
      added_at:
        type: string
      artist:
        type: string
      created_at:
        type: string
      duration:
        description: secondes
        type: integer
      file_url:
        description: URL Supabase Storage
        type: string
      id:
        type: string
      playlist_id:
        description: Champs de liaison playlist (remplis quand la track est dans une
          playlist)
        type: string
      position:
        type: integer
      title:
        type: string
      uploaded_by:
        type: string
    type: object
...
  handler.AddFavoriteRequest:
    properties:
      stream_id:
        type: string
    required:
    - stream_id
    type: object
...
host: localhost:8080
info:
  contact:
    name: StreamPulse team
    url: https://github.com/Ghost-15/streaming
  description: |-
    Live audio streaming API. A broadcaster ingests an audio
    stream, the Hub fans it out to every listener, and playlists,
    favorites and listening history are exposed as REST resources.
  license:
    name: Educational project
  termsOfService: https://github.com/Ghost-15/streaming
  title: StreamPulse API
  version: "1.0"
paths:
...
  /api/v1/favorites:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/entity.Track'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the favorites of the authenticated user
      tags:
      - favorites
    post:
      consumes:
      - application/json
      parameters:
      - description: Favorite payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AddFavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a live stream to the favorites
      tags:
      - favorites
  /api/v1/favorites/{streamID}:
    delete:
      parameters:
      - description: Stream ID
        in: path
        name: streamID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a live stream from the favorites
      tags:
      - favorites
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Access token issued by /api/v1/auth/login, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"